The change of messages model to inherit board fk
